Move viewer/format/*.py into viewer/ to prevent duplicate with built-in function: format
