--- microproxy/viewer/format/__init__.py
@@ -1,1 +1,0 @@
-from formatter import Formatter
